manage res files
