+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,14 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/xmlFragment">
 
     <fragment
-        android:id="@+id/FirstFragment"
+        android:id="@+id/xmlFragment"
         android:name="com.konovalov.draggable.coin.view.xml.XmlFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_xml">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
+        android:label="XmlFragment" />
     <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.konovalov.draggable.coin.view.compose.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_xml">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
+        android:id="@+id/composeFragment"
+        android:name="com.konovalov.draggable.coin.view.compose.ComposeFragment"
+        android:label="ComposeFragment" />
 </navigation>
