+++ core/common/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
