+++ core/common/src/main/res/mipmap-mdpi/ic_launcher_round.webp
