+++ core/common/src/main/res/values/strings.xml
