+++ core/common/src/main/res/xml/backup_rules.xml
