+++ core/ui/src/main/java/com/konovalov/draggable/coin/view/core/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.konovalov.draggable.coin.view.ui.theme
+package com.konovalov.draggable.coin.view.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
