+++ core/ui/src/main/java/com/konovalov/draggable/coin/view/core/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.konovalov.draggable.coin.view.ui.theme
+package com.konovalov.draggable.coin.view.core.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
