+++ features/compose/src/main/AndroidManifest.xml
@@ -1,11 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest>
-    <application>
-        <activity
-            android:name=".ComposeActivity"
-            android:exported="false"
-            android:label="@string/title_activity_compose"
-            android:theme="@style/Theme.DraggableScaredView" />
-    </application>
-
-</manifest>
+<manifest/>
