+++ features/compose/src/main/java/com/konovalov/draggable/coin/view/compose/ComposeFragment.kt
@@ -1,0 +1,40 @@
+package com.konovalov.draggable.coin.view.compose
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.tooling.preview.Preview
+import com.konovalov.draggable.coin.view.core.ui.theme.DraggableScaredViewTheme
+
+class ComposeFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ) = ComposeView(requireContext()).apply {
+        setContent {
+            ComposeFragmentHolder()
+        }
+    }
+}
+
+@Preview
+@Composable
+fun ComposeFragmentHolder() =
+    DraggableScaredViewTheme {
+        Surface(
+            modifier = Modifier.fillMaxSize(),
+            color = MaterialTheme.colors.background
+        ) {
+            Greeting("Android")
+            Text(text = "Hello world.")
+        }
+    }
