--- features/compose/src/main/res/values-land/dimens.xml
@@ -1,3 +1,0 @@
-<resources>
-    <dimen name="fab_margin">48dp</dimen>
-</resources>
